made a few changes so that it doesn't deletes message when someone misses a field

+++ commands/donate.js
@@ -20,7 +20,7 @@ module.exports = {
         const msg = fields[4];
 
         if (message.channel.id !== '738458175187058698') return;
-        if (!prize || !time || !winners || !req || !msg || args[0] === 'help') {
+        if (args[0] === 'help') {
             const format = new Discord.MessageEmbed()
                 .setTitle(`Donate Command`)
                 .setDescription(`Use the \`-donate\` command to donate for giveaways and ping giveaway managers\n - There's a 10 minute cooldown on the command to avoid abuse and spam`)
@@ -30,22 +30,26 @@ module.exports = {
             message.delete().then
             message.channel.send(format);
         } else
-            if (cooldown) {
-                const remaining = Duration(cooldown - Date.now(), { units: ['m', 's'], round: true });
-                return message.channel.send(`You just used this command \nYou need to wait **${remaining}** before using it again.`).catch((err) => msg.reply(`${err}`));
-            }
-            else {
-                const donate = new Discord.MessageEmbed()
-                    .setTitle(`Thanks for the donation`)
-                    .addFields({ name: `Donator`, value: `${message.author} (${message.author.tag}) ` }, { name: 'Prize', value: prize }, { name: 'Time', value: time }, { name: 'Winners', value: winners }, { name: "Requirement", value: req }, { name: "Message", value: msg })
-                    .setThumbnail(message.author.displayAvatarURL({ dynamic: true }))
-                    .setColor(message.member.displayHexColor)
-                    .setFooter(`wait patiently for a giveaway manager to reply!`);
-                message.delete().then
-                message.channel.send({ content: `<@&688386729807577284>`, embed: donate });
-                lastUsed.set(message.author.id, Date.now() + 1000 * 60 * 10);
-                setTimeout(() => { lastUsed.delete(message.author.id) }, 1000 * 60 * 10);
-            }
+            if (!prize || !time || !winners || !req || !msg) {
+                message.channel.send(`Correct Format to use is \`-donate <prize>, <time>, <winners>, <requirements>, <message>\`\n - Don't leave any fields empty.`);
+                return;
+            } else
+                if (cooldown) {
+                    const remaining = Duration(cooldown - Date.now(), { units: ['m', 's'], round: true });
+                    return message.channel.send(`You just used this command \nYou need to wait **${remaining}** before using it again.`).catch((err) => msg.reply(`${err}`));
+                }
+                else {
+                    const donate = new Discord.MessageEmbed()
+                        .setTitle(`Thanks for the donation`)
+                        .addFields({ name: `Donator`, value: `${message.author} (${message.author.tag}) ` }, { name: 'Prize', value: prize }, { name: 'Time', value: time }, { name: 'Winners', value: winners }, { name: "Requirement", value: req }, { name: "Message", value: msg })
+                        .setThumbnail(message.author.displayAvatarURL({ dynamic: true }))
+                        .setColor(message.member.displayHexColor)
+                        .setFooter(`wait patiently for a giveaway manager to reply!`);
+                    message.delete().then
+                    message.channel.send({ content: `<@&688386729807577284>`, embed: donate });
+                    lastUsed.set(message.author.id, Date.now() + 1000 * 60 * 10);
+                    setTimeout(() => { lastUsed.delete(message.author.id) }, 1000 * 60 * 10);
+                }
 
     }
